fix(language): keep object literal CI paths green
