+++ crates/perry-runtime/src/object/mod.rs
@@ -46,6 +46,7 @@ mod native_call_method;
 mod native_module;
 mod native_module_dispatch;
 mod native_module_stream;
+mod object_literal_ops;
 mod object_ops;
 mod object_ops_frozen;
 mod polymorphic_index;
@@ -79,6 +80,7 @@ pub use native_call_method::*;
 pub use native_module::*;
 pub(crate) use native_module_dispatch::*;
 pub(crate) use native_module_stream::*;
+pub use object_literal_ops::*;
 pub use object_ops::*;
 pub use object_ops_frozen::*;
 pub use polymorphic_index::*;
